+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2025-01-07]
+## [0.0.1 - 2025-01-20]
 ### Added
 - First version of the project
 - Spring Application
@@ -185,3 +185,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Send error email if error while sending query to bridgehead
 - Mailing Black List
 - Sort Mailing Black List by first and last name
+- Results
+- DRN in exporter label
+- Final User State in results
+- Create State in Frontend DTO's (for project and project bridgehead)
+- Fetch email message and subject HTTP endpoint
+- Send project results only if there are final users
+- User Project State in State Constraints